Refactored PNLReport by separating hooks into a custom hook named usePnlStates and storing the pre-existing hook usePnlReport in the same file, also exported all the utilities to the pnlReport.utils file.
